backend/requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.22.1
Faker==12.0.1
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.1
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.2.2
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.9
